оптимизация кода и усовершениствование функции валидации
